trampoline & switch

+++ src/Hitbox.java
@@ -5,7 +5,7 @@ public class Hitbox {
 	public double[] position;
 	public double[] speed;
 	public int nbPoints;
-	public short type; 		
+	public int type; 		
 	/* 0: not visible, 1: solid, 2: plateform */
 	
 	protected double s;
